requirements.txt
annotated-types==0.7.0
anthropic==0.66.0
anyio==4.10.0
certifi==2025.8.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
MouseInfo==0.1.3
pillow==11.3.0
PyAutoGUI==0.9.54
# macOS-only deps (guard with markers to avoid Windows/Linux install)
PyObjC==11.1; sys_platform == "darwin"
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.1; sys_platform == "darwin"
pyobjc-framework-Cocoa==11.1; sys_platform == "darwin"
pyobjc-framework-Quartz==11.1; sys_platform == "darwin"
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
rubicon-objc==0.5.2; sys_platform == "darwin"
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
pytest==8.3.3
pytest-timeout==2.3.1
injector==0.21.0
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
orjson==3.10.7
pystray==0.19.5
Pillow>=11.3.0
